fix(ConnectionMock.spec.js): monkey-patch with PromiseMock for these tests
Previously, tests were failing due to depending half on patched code, and half on native promises
which did not interact with the mocks as expected. Calling PromiseBackend.patchWithMock() before
each test, and restoring the native promise after each test, corrects this behaviour. Ideally,
this should occur for all test files, but for now the problem is limited to ConnectionMock.spec.js.

Closes #18
Closes #19